common: move Windows files to a separate folder
We ended up with quite a few files having the "_win32" suffix.
It's probably better if we move them to a separate folder and drop
the suffix.
