+++ src/common/win32/dlfcn.cc
