+++ src/common/win32/errno.cc
