+++ src/common/win32/registry.h
