+++ src/common/win32/service.h
